fix: rank asymmetric peg warnings
